+++ src/Cwaff-Guns/Deadline.cs
@@ -28,7 +28,7 @@ public class Deadline : CwaffGun
             fireAudio: "deadline_fire_sound", reloadAudio: "deadline_reload_sound")
           .AddToShop(ModdedShopType.TimeTrader)
           .AddToShop(ModdedShopType.Boomhildr)
-          .InitProjectile(GunData.New(clipSize: 8, cooldown: 0.4f, angleVariance: 0.0f, shootStyle: ShootStyle.SemiAutomatic,
+          .InitProjectile(GunData.New(clipSize: 8, cooldown: 0.4f, angleVariance: 0.0f, shootStyle: ShootStyle.SemiAutomatic, customClip: true,
             speed: 60.0f, range: 30.0f, sprite: "deadline_projectile", fps: 2, anchor: Anchor.MiddleLeft, collidesWithEnemies: false))
           .Attach<DeadlineProjectile>()
           .Attach<EasyTrailBullet>(trail => {